Pin numpy version and update .gitignore
