+++ setup.cfg
@@ -6,7 +6,7 @@ max-line-length = 79
 ignore = D202,D203,W503,E203, E501, F841, E731
 
 [isort]
-known_third_party = celery,dotenv,ee,geojson,geopandas,geoviews,httplib2,imageio,initials,matplotlib,metar,models,netCDF4,numpy,pandas,psycopg2,pydrive,rasterio,rasterstats,requests,responses,setuptools,shapely,sqlalchemy,xarray,yaml
+known_third_party = celery,dotenv,ee,fill_counties,geojson,geopandas,geoviews,httplib2,imageio,initials,matplotlib,metar,models,netCDF4,numpy,pandas,psycopg2,pydrive,rasterio,rasterstats,requests,responses,setuptools,shapely,sqlalchemy,xarray,yaml
 ensure_newline_before_comments=true
 line_length = 79
 multi_line_output = 3
