+++ utilities/fill_counties_api.py
@@ -1,16 +1,12 @@
 import csv
 import os
-import time
-import warnings
 from multiprocessing.pool import Pool
 from os.path import abspath, dirname
 from os.path import join as join_path
 
-import geojson
 import psycopg2
-import requests
 from dotenv import load_dotenv
-from initials import initials
+from fill_counties import to_row
 
 load_dotenv()
 
@@ -21,59 +17,6 @@ db_config = {
     "host": os.getenv("PSQL_HOST"),
     "port": os.getenv("PSQL_PORT"),
 }
-
-
-# Get geojson from geocode
-def get_api(geocodigo):
-    return f"https://servicodados.ibge.gov.br/api/v3/malhas/municipios/{geocodigo}?formato=application/vnd.geo+json"
-
-
-# Return data from API
-def read_json(geocodigo):
-    url = get_api(geocodigo)
-    status = 0
-    wait = 3
-    while status != 200 and wait <= 16:
-        resp = requests.get(url)
-        status = resp.status_code
-        time.sleep(wait)
-        wait *= 3
-    resp_data = resp.json()
-    return resp_data
-
-
-# Parse data
-# Compare cities(Nome_Município) in the csv file
-def county_polygon(county_code):
-    for feature in read_json(county_code)['features']:
-        if feature["properties"].get("codarea") == county_code:
-            return geojson.dumps(feature['geometry'])
-    warnings.warn("is not in this geojson: {}.".format(county_code))
-    return
-
-
-# Prepare dict to save in the database
-def to_row(county):
-    county_code = county["Cod Municipio Completo"]
-    name = county["Nome_Município"]
-    uf = county["Nome_UF"]
-    try:
-        geojson = county_polygon(county_code)
-        print(county_code, "|", name, "|", uf)
-    except ValueError as e:
-        print(e)
-        geojson = ""
-
-    return dict(
-        county_code=county_code,
-        name=name,
-        geojson=geojson,
-        uf=initials[uf],
-        population=0,
-    )
-
-
-#
 
 
 def to_save(data, schema="Dengue_global", table="Municipio"):
@@ -114,6 +57,6 @@ def to_save(data, schema="Dengue_global", table="Municipio"):
 
 
 BASE_DIR = dirname(abspath(__file__))
-path = join_path(BASE_DIR, "DTB_2021_Municipio_faltantes.csv")
+path = join_path(BASE_DIR, "DTB_Only_Municipio_faltantes.csv")
 rows = Pool().map(to_row, csv.DictReader(open(path)))
 to_save(rows, schema="Dengue_global", table="Municipio")
